requirements.txt
Flask==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.5.1
flask-cors==4.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Flask-OAuthlib==0.9.6
PyJWT==2.10.1
pytest==8.3.4
pytest-flask==1.3.0
requests==2.32.3
google-auth==2.37.0
google-auth-oauthlib==1.2.1
python-socketio==5.12.1
python-engineio==4.11.2
